% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_aggregate_raster.R
\name{aggregate_raster}
\alias{aggregate_raster}
\title{Aggregate raster layers into fixed groups}
\usage{
aggregate_raster(x, n, fun = mean, cores = 1)
}
\arguments{
\item{x}{A \code{terra::SpatRaster} with multiple layers.}

\item{n}{Integer scalar giving the number of layers to combine in each group.}

\item{fun}{Function used to aggregate layers within each group. Defaults to
\code{mean}.}

\item{cores}{Integer number of CPU cores to use. Defaults to \code{1}.}
}
\value{
A \code{terra::SpatRaster} with aggregated layers.
}
\description{
Aggregates consecutive layers of a multilayer \code{terra::SpatRaster} into
groups of size \code{n} using a summary function.
}
\details{
Layers are grouped in sequential blocks of size \code{n}. If the number of layers
in \code{x} is not an exact multiple of \code{n}, the final group contains the remaining
layers.

Aggregation is performed with \code{terra::tapp()}.
}
\examples{
\dontrun{
library(terra)

r <- rast(nrows = 10, ncols = 10, nlyrs = 12)
values(r) <- runif(ncell(r) * nlyr(r))

out <- aggregate_raster(r, n = 3, fun = mean)
out
}

}
