% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclim_dataprep.R
\name{checkinputs}
\alias{checkinputs}
\title{Check format and values in model inputs}
\usage{
checkinputs(weather, vegp, soilc, dtm, windhgt = 2)
}
\arguments{
\item{weather}{a data.frame of weather variables (see details)}

\item{vegp}{an object of class vegparams as returned by \code{\link[=vegpfromhab]{vegpfromhab()}} (see details)}

\item{soilc}{an object of class soilcharac as returned by \code{\link[=soilcfromtype]{soilcfromtype()}}}

\item{dtm}{a SpatRaster object of elevations (see details)}

\item{windhgt}{height above ground (m) of wind speed measurement}

\item{precip}{a vector of daily precipitation}

\item{daily}{optional logical indicating whether input weather data are daily}

\item{tstep}{one of \code{hour} or \code{day} indicating whether data are hourly or daily}
}
\value{
a list of checked inputs
}
\description{
The function \code{checkinputs} checks all the inputs into the model and returns errors
and warnings
}
\details{
The format and and units of \code{weather} must follow that in the example
dataset \code{climdata}. The array of Plant Area index values in \code{vegp} must
of the same x and y dims as \code{dtm} but can contain any number of repeated
measures up to the number of entries in \code{weather}. Data are interpolated to the
time increment of \code{weather}. Other vegetation paramaters, including vegetation
height are assumed time-invarient. The SpatRaster datasets in \code{soilc} must have
the same x and y dims as \code{dtm}. The x,y and z units of \code{dtm} must be all be in
metres and the coordinate reference system must be defined.
}
\examples{
# No warnings or errors given:
checks<-checkinputs(climdata, vegp, soilc, dtmcaerth)
# Warning given (not run)
# weather<-climdata
# weather$relhum[1]<-101
# checks<-checkinputs(weather, vegp, soilc, dtmcaerth)
# Error given (NB not run)
# weather<-climdata
# weather$pres<-weather$pres*1000
# checks<-checkinputs(weather, vegp, soilc, dtmcaerth)
# Error given for vegp (not run)
# vegp2<-vegp
# vegp2$clump<-1
# checks<-checkinputs(climdata, vegp2, soilc, dtmcaerth)
# Warning given for vegp (not run)
# vegp2<-vegp
# vegp2$pai<-vegp$pai*10
# checks<-checkinputs(climdata, vegp2, soilc, dtmcaerth)
}
