% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco_days_since_peak_IRG.R
\name{days_since_peak_IRG}
\alias{days_since_peak_IRG}
\title{Days since peak IRG}
\usage{
days_since_peak_IRG(x, absolute_value = F)
}
\arguments{
\item{x}{A \code{terra::SpatRaster} with multiple layers ordered in time.}

\item{absolute_value}{Logical. If \code{TRUE}, returns the absolute value of
the output (i.e., distance in time from peak regardless of direction).}
}
\value{
A \code{terra::SpatRaster} with the same number of layers as \code{x}.
Values are negative before the peak, \code{0} at the peak, and positive after
the peak unless \code{absolute_value = TRUE}.
}
\description{
Calculates, for each layer in a multilayer IRG raster, the number of
time steps since the pixel-specific peak IRG value occurred.
}
\details{
The peak layer is identified separately for each cell using
\code{terra::which.max()}. The output is calculated as:

\code{layer_index - peak_index}

If \code{absolute_value = TRUE}, the result is transformed to:

\code{abs(layer_index - peak_index)}
}
\examples{
\dontrun{
library(terra)

r <- rast(nrows = 10, ncols = 10, nlyrs = 5)
values(r) <- runif(ncell(r) * nlyr(r))

out <- days_since_peak_IRG(r)
out_abs <- days_since_peak_IRG(r, absolute_value = TRUE)
}

}
