% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclim_internal.R
\name{.biomicropoint}
\alias{.biomicropoint}
\title{Run micropoint for bioclim calculations}
\usage{
.biomicropoint(weather, tme, reqhgt, vegp, soilc, dtm, zref, windhgt, soilm)
}
\description{
Run micropoint for bioclim calculations
}
