% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrition_print.R
\name{format.fresh_output}
\alias{format.fresh_output}
\title{Format method for fresh_output objects}
\usage{
\method{format}{fresh_output}(x, ...)
}
\arguments{
\item{x}{A \code{fresh_output} object.}

\item{...}{Not used.}
}
\description{
Controls the description displayed in the RStudio Environment pane.
}
