% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palettes.R
\name{pal_continuous_n}
\alias{pal_continuous_n}
\title{Get interpolated continuous palette}
\usage{
pal_continuous_n(n = 256)
}
\arguments{
\item{n}{Number of colors to interpolate.}
}
\value{
A character vector of \code{n} hex colors.
}
\description{
Returns \code{n} interpolated colors from the \code{"cont"} gradient.
}
