% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrition_print.R
\name{print.fresh_output}
\alias{print.fresh_output}
\title{Print method for fresh_output objects}
\usage{
\method{print}{fresh_output}(x, ...)
}
\arguments{
\item{x}{A \code{fresh_output} object.}

\item{...}{Not used.}
}
\description{
Controls console output when a \code{fresh_output} object is typed or
explicitly printed.
}
