% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclim_Rimplementation.R
\name{soiltemp}
\alias{soiltemp}
\title{Calculates ground surface temperature (hourly)}
\usage{
soiltemp(micro, reqhgt = 0.05, pai_a = NA, tfact = 1.5)
}
\arguments{
\item{micro}{an object of class \code{microin} as returned by \code{\link[=modelin]{modelin()}} (see details)}

\item{reqhgt}{height above ground at which model outputs are needed (m).}

\item{pai_a}{plant area index above \code{reqhgt}. Determined from total \code{pai} if not supplied.}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see \code{\link[=soilmdistribute]{soilmdistribute()}})}
}
\value{
an object of class micro with the following components added:
\describe{
\item{T0}{ground surface temperature (deg C)}
\item{G}{ground heat flux (W/m^2)}
\item{...}{additional terms needed for subsequent modelling}
}
}
\description{
The function \code{soiltemp} estimates ground surface temperature
}
\details{
if \code{\link[=wind]{wind()}} has not been run to add additional elements to \code{micro}
it is automatically called.
}
\seealso{
\code{\link[=wind]{wind()}}
}
