% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclim_dataprep.R
\name{subsetpointmodela}
\alias{subsetpointmodela}
\title{Subsets outputs from point microclimate model run over arrays}
\usage{
subsetpointmodela(pointmodela, tstep = "month", what = "tmax", days = NA)
}
\arguments{
\item{pointmodela}{a list of model outputs from the point microclimate model applied over arrays as returned by \code{\link[=runpointmodela]{runpointmodela()}}.}

\item{tstep}{one of \code{year} or \code{month} (see details)}

\item{what}{one of \code{tmax}, \code{tmin} or \code{tmedian} (maximum, minimum or median temperature repsectively - see details)}

\item{days}{optionally a vector of the days in the time sequence  to return data for (if provided tstep is ignored)}
}
\value{
A list with the same format as \code{pointmodela} but with specified values
'only selected.
}
\description{
The function \code{subsetpointmodela} is the equivalent of \code{\link[=subsetpointmodel]{subsetpointmodel()}} used
for outputs from \code{\link[=`runpointmodela]{`runpointmodela()}}
}
\seealso{
\code{\link[=subsetpointmodel]{subsetpointmodel()}}, \code{\link[=runpointmodela]{runpointmodela()}}
}
