% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclim_Rimplementation.R
\name{twostream}
\alias{twostream}
\title{Applies two-stream radiation model}
\usage{
twostream(micro, reqhgt = 0.05, pai_a = NA, tfact = 1.5)
}
\arguments{
\item{micro}{an object of class \code{micro} as returned by \code{\link[=modelin]{modelin()}}}

\item{reqhgt}{height at which temperatures are required (m). Negative if below ground}

\item{pai_a}{a SpatRaster of plant area index values above \code{reqhgt}. Determined from
total \code{pai} if not supplied. Must patch the dimensions of \code{vegp$pai} if supplied.}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation}
}
\value{
an object of class micro with additional terms added for subsequent modelling
}
\description{
The function \code{twostream} applies a varient of the Sellers two-stream
radiation model to calaculate long and showtwave radiation absorbed by the ground and
canopy and shortwave radiation absorbed by leafs.
}
