% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclim_Rimplementation.R
\name{wind}
\alias{wind}
\title{Downscales wind}
\usage{
wind(micro, reqhgt = 0.05, pai_a = NA, tfact = 1.5)
}
\arguments{
\item{micro}{object of class microin as returned by \code{\link[=modelin]{modelin()}}}

\item{reqhgt}{height above ground for which wind speeds are wanted. If negative (below ground) wind friction velocity only is returned}

\item{pai_a}{plant area index above \code{reqhgt}. Determined from total \code{pai} if not supplied.}

\item{tfact}{coefficient determining sensitivity of soil moisture to variation
in topographic wetness (see \code{\link[=soilmdistribute]{soilmdistribute()}})}
}
\value{
an object of class microin with the following components added:
\describe{
\item{uf}{wind friction velocity (m/s)}
\item{uz}{if \code{reqhgt > 0} wind speed at height \code{reqhgt} (m/s)}
\item{gHa}{convective conductance (mol/m^2/s)}
}
}
\description{
The function \code{wind} downscales wind speed accounting for
vegetation and terrain
}
\details{
In downscaling wind, two processes are accounted for. Firstly the drag effects
of vegetation on wind , which ultimately dictate the wind height profile. Secondly,
the effects of local vegetation and terrain on wind speed. Terrain effects are calculated
by applying the topographic shelter coefficient described in Maclean et al (2019) Methods
in Ecology and Evolution, 10:280-290.
}
