% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrml.R
\name{vrml}
\alias{vrml}
\title{Vector ruggedness of local relief}
\usage{
vrml(x, s)
}
\arguments{
\item{x}{A \code{terra::SpatRaster} DEM.}

\item{s}{Odd integer giving the focal window size.}
}
\value{
A \code{terra::SpatRaster} of vector ruggedness values.
}
\description{
Calculates a smoothed DEM, subtracts the original DEM to isolate local
topographic variation, and then computes vector ruggedness measure on the
residual surface.
}
\details{
The DEM is first smoothed with \code{terra::focal()} using a mean filter of size
\code{s}. The smoothed surface is then subtracted from the original DEM to isolate
local topographic variation. Vector ruggedness measure is then calculated on
that residual raster using \code{spatialEco::vrm()}.
}
\examples{
\dontrun{
library(terra)

r <- rast(nrows = 100, ncols = 100, xmin = 0, xmax = 100, ymin = 0, ymax = 100)

xy <- as.data.frame(crds(r))
z <- with(
  xy,
  800 +
    250 * exp(-((x - 30)^2 + (y - 35)^2) / 250) +
    180 * exp(-((x - 70)^2 + (y - 65)^2) / 180) -
    120 * exp(-((x - 55)^2 + (y - 45)^2) / 120) +
    60 * sin(x / 8) * cos(y / 10)
)

values(r) <- z

out <- vrml(r, s = 5)
plot(out)
}

}
